#6 ロジックをhooksのみ分離して記述（ImageUpload以外）
