+++ src/components/pages/english/EnglishWordPracPage.tsx
@@ -3,15 +3,19 @@
 import { Box, Stack } from '@mui/material';
 import React from 'react';
 import { PrintInfo } from '@/components/domains/english/word_prac/PrintInfo';
-import { SideSectionList } from '@/components/domains/english/word_prac/SideSectionList';
+import { SideSessionList } from '@/components/domains/english/word_prac/SideSessionList';
 import { WordList } from '@/components/domains/english/word_prac/WordList';
+import { useEnglishWordPrac } from '@/components/hooks/english/useEnglishWordPrac';
 
-export const EnglishWordPracPage: React.FC = () => (
-  <Box display="flex">
-    <SideSectionList />
-    <Box p={2} component={Stack} spacing={2}>
-      <PrintInfo />
-      <WordList />
+export const EnglishWordPracPage: React.FC = () => {
+  const englishWordPrac = useEnglishWordPrac();
+  return (
+    <Box display="flex">
+      <SideSessionList sessions={englishWordPrac.sessions} />
+      <Box p={2} component={Stack} spacing={2}>
+        <PrintInfo />
+        <WordList words={englishWordPrac.words} />
+      </Box>
     </Box>
-  </Box>
-);
+  );
+};
